+++ src/main/java/org/folio/roles/integration/kafka/model/CapabilityEvent.java
@@ -1,9 +1,12 @@
 package org.folio.roles.integration.kafka.model;
 
+import com.fasterxml.jackson.annotation.JsonIgnore;
 import java.util.List;
+import java.util.Map;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import org.folio.common.utils.permission.model.PermissionData;
 
 @Data
 @NoArgsConstructor
@@ -29,6 +32,12 @@ public class CapabilityEvent {
    * List with folio resources (permission and corresponding endpoints).
    */
   private List<FolioResource> resources;
+
+  /**
+   * Permission data mapping overrides.
+   */
+  @JsonIgnore
+  private Map<String, PermissionData> permissionMappingOverrides;
 
   /**
    * Sets moduleId field and returns {@link CapabilityEvent}.
@@ -67,6 +76,16 @@ public class CapabilityEvent {
    */
   public CapabilityEvent resources(List<FolioResource> resources) {
     this.resources = resources;
+    return this;
+  }
+
+  /**
+   * Sets permissionMappingOverrides field and returns {@link CapabilityEvent}.
+   *
+   * @return modified {@link CapabilityEvent} value
+   */
+  public CapabilityEvent permissionMappingOverrides(Map<String, PermissionData> permissionMappingOverrides) {
+    this.permissionMappingOverrides = permissionMappingOverrides;
     return this;
   }
 }